feat: 출고 지연 기능 지원을 위한 DB 스키마 및 엔티티 수정
Shipments 엔티티와 데이터베이스 스키마 간의 불일치 문제를 해결하고, 판매자 배송 관리 기능 중 '출고 지연 요청' 기능을 지원하기 위해 전체적으로 수정했습니다.

프로젝트 요구사항 분석 결과, `expected_ship_date`와 `delay_reason` 필드는 판매자의 출고 지연 처리 기능에 필수적이었습니다. 따라서 단순히 엔티티에서 필드를 제거하는 대신, 데이터베이스 스키마에 해당 컬럼들을 추가하여 문제를 해결했습니다.

**주요 변경 사항:**

1.  **DB 스키마 수정 (`schema.sql`)**
    * `shipments` 테이블에 `expected_ship_date` (TIMESTAMP) 컬럼 추가
    * `shipments` 테이블에 `delay_reason` (VARCHAR) 컬럼 추가

2.  **Shipments 엔티티 리팩토링 (`Shipments.java`)**
    * DB에 추가된 컬럼과 매핑되는 `expectedShipDate`, `delayReason` 필드를 유지했습니다.
    * 서비스 레이어에서 직접 `setter`를 사용하는 패턴이 확인되어, 중복되고 사용되지 않는 아래의 편의 메서드들을 제거하여 코드 명확성을 높였습니다.
        * `isShipped()`, `setTrackingInfo()`, `markAsDelivered()`, `hideFromSeller()`, `showToSeller()`, `updateShipmentMemo()`
    * Builder 패턴과 기존 코드 호환성을 위한 별칭 메서드들은 그대로 유지하여 유연성과 안정성을 확보했습니다.

**기대 효과:**
* 데이터베이스와 엔티티 간의 매핑 오류를 완전히 해결했습니다.
* 판매자용 '출고 지연 요청' 기능의 백엔드 지원을 완료했습니다.
* 불필요한 코드를 제거하여 엔티티 클래스의 유지보수성을 향상시켰습니다.

+++ src/main/java/com/team5/catdogeats/orders/service/impl/SellerOrderServiceImpl.java
@@ -17,16 +17,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 /**
- * 판매자용 주문 관리 서비스 구현체 (CQRS Facade 패턴)
- * 기존 인터페이스를 유지하면서 내부적으로 CQRS 패턴으로 재구성된 서비스들을 조합
- *
- * CQRS 구조:
- * - SellerOrderQueryService: 읽기 전용 작업 (주문 조회 관련)
- * - SellerOrderCommandService: 쓰기 전용 작업 (주문 변경 관련)
- *
- * 기존 5개 서비스에서 2개 CQRS 서비스로 재구성:
- * [기존] SellerOrderQueryService + SellerOrderListService → [신규] SellerOrderQueryService (읽기)
- * [기존] SellerOrderStatusService + SellerOrderTrackingService + SellerOrderVisibilityService → [신규] SellerOrderCommandService (쓰기)
+ * 판매자용 주문 관리 서비스 구현체
  */
 @Slf4j
 @Service
@@ -138,7 +129,7 @@ public class SellerOrderServiceImpl implements SellerOrderService {
                 .orders(filteredOrders)
                 .currentPage(response.currentPage())
                 .totalPages(response.totalPages())
-                .totalElements((long) filteredOrders.size())
+                .totalElements(filteredOrders.size())
                 .pageSize(response.pageSize())
                 .hasNext(response.hasNext())
                 .hasPrevious(response.hasPrevious())
@@ -166,7 +157,7 @@ public class SellerOrderServiceImpl implements SellerOrderService {
                 .orders(filteredOrders)
                 .currentPage(response.currentPage())
                 .totalPages(response.totalPages())
-                .totalElements((long) filteredOrders.size())
+                .totalElements(filteredOrders.size())
                 .pageSize(response.pageSize())
                 .hasNext(response.hasNext())
                 .hasPrevious(response.hasPrevious())
